requirements.txt
boto3==1.34.15
botocore==1.34.15
python-dotenv==1.0.0
python-twitter-v2==0.9.0
requests==2.31.0
pytz~=2023.3.post1
PyYAML~=6.0.1